multi_agent_email/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
email-validator
chromadb
requests